+++ src/oauth_codex/resources/responses/responses.py
@@ -2,31 +2,16 @@ from __future__ import annotations
 
 from typing import Any, AsyncIterator, Iterator, Literal, overload
 
-from ...legacy_types import (
-    Message,
-    ToolInput,
-    ToolResult,
-    TruncationMode,
-    ValidationMode,
-)
 from ..._resource import AsyncAPIResource, SyncAPIResource
-from .._unsupported import raise_unsupported
+from ...core_types import Message, ToolInput, ToolResult, TruncationMode, ValidationMode
+from ...types.responses import Response, ResponseStreamEvent
 from .._wrappers import (
     async_to_raw_response_wrapper,
     async_to_streamed_response_wrapper,
     to_raw_response_wrapper,
     to_streamed_response_wrapper,
 )
-from ...types.responses import Response, ResponseStreamEvent
-from ._helpers import aiter_events, iter_events, response_from_legacy
-from .input_items import (
-    AsyncInputItems,
-    AsyncInputItemsWithRawResponse,
-    AsyncInputItemsWithStreamingResponse,
-    InputItems,
-    InputItemsWithRawResponse,
-    InputItemsWithStreamingResponse,
-)
+from ._helpers import aiter_engine_events, iter_engine_events, response_from_engine
 from .input_tokens import (
     AsyncInputTokens,
     AsyncInputTokensWithRawResponse,
@@ -38,10 +23,6 @@ from .input_tokens import (
 
 
 class Responses(SyncAPIResource):
-    @property
-    def input_items(self) -> InputItems:
-        return InputItems(self._client)
-
     @property
     def input_tokens(self) -> InputTokens:
         return InputTokens(self._client)
@@ -175,8 +156,8 @@ class Responses(SyncAPIResource):
             **extra,
         )
         if stream:
-            return iter_events(out)
-        return response_from_legacy(out)
+            return iter_engine_events(out)
+        return response_from_engine(out)
 
     def stream(
         self,
@@ -238,21 +219,6 @@ class Responses(SyncAPIResource):
             **extra,
         )
 
-    def retrieve(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.retrieve")
-
-    def cancel(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.cancel")
-
-    def delete(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.delete")
-
-    def compact(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.compact")
-
-    def parse(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.parse")
-
     @property
     def with_raw_response(self) -> ResponsesWithRawResponse:
         return ResponsesWithRawResponse(self)
@@ -263,10 +229,6 @@ class Responses(SyncAPIResource):
 
 
 class AsyncResponses(AsyncAPIResource):
-    @property
-    def input_items(self) -> AsyncInputItems:
-        return AsyncInputItems(self._client)
-
     @property
     def input_tokens(self) -> AsyncInputTokens:
         return AsyncInputTokens(self._client)
@@ -400,8 +362,8 @@ class AsyncResponses(AsyncAPIResource):
             **extra,
         )
         if stream:
-            return aiter_events(out)
-        return response_from_legacy(out)
+            return aiter_engine_events(out)
+        return response_from_engine(out)
 
     async def stream(
         self,
@@ -463,21 +425,6 @@ class AsyncResponses(AsyncAPIResource):
             **extra,
         )
 
-    async def retrieve(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.retrieve")
-
-    async def cancel(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.cancel")
-
-    async def delete(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.delete")
-
-    async def compact(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.compact")
-
-    async def parse(self, *_: Any, **__: Any) -> None:
-        raise_unsupported("responses.parse")
-
     @property
     def with_raw_response(self) -> AsyncResponsesWithRawResponse:
         return AsyncResponsesWithRawResponse(self)
@@ -491,12 +438,6 @@ class ResponsesWithRawResponse:
     def __init__(self, responses: Responses) -> None:
         self.create = to_raw_response_wrapper(responses.create)
         self.stream = to_raw_response_wrapper(responses.stream)
-        self.retrieve = to_raw_response_wrapper(responses.retrieve)
-        self.cancel = to_raw_response_wrapper(responses.cancel)
-        self.delete = to_raw_response_wrapper(responses.delete)
-        self.compact = to_raw_response_wrapper(responses.compact)
-        self.parse = to_raw_response_wrapper(responses.parse)
-        self.input_items = InputItemsWithRawResponse(responses.input_items)
         self.input_tokens = InputTokensWithRawResponse(responses.input_tokens)
 
 
@@ -504,12 +445,6 @@ class AsyncResponsesWithRawResponse:
     def __init__(self, responses: AsyncResponses) -> None:
         self.create = async_to_raw_response_wrapper(responses.create)
         self.stream = async_to_raw_response_wrapper(responses.stream)
-        self.retrieve = async_to_raw_response_wrapper(responses.retrieve)
-        self.cancel = async_to_raw_response_wrapper(responses.cancel)
-        self.delete = async_to_raw_response_wrapper(responses.delete)
-        self.compact = async_to_raw_response_wrapper(responses.compact)
-        self.parse = async_to_raw_response_wrapper(responses.parse)
-        self.input_items = AsyncInputItemsWithRawResponse(responses.input_items)
         self.input_tokens = AsyncInputTokensWithRawResponse(responses.input_tokens)
 
 
@@ -517,12 +452,6 @@ class ResponsesWithStreamingResponse:
     def __init__(self, responses: Responses) -> None:
         self.create = to_streamed_response_wrapper(responses.create)
         self.stream = to_streamed_response_wrapper(responses.stream)
-        self.retrieve = to_streamed_response_wrapper(responses.retrieve)
-        self.cancel = to_streamed_response_wrapper(responses.cancel)
-        self.delete = to_streamed_response_wrapper(responses.delete)
-        self.compact = to_streamed_response_wrapper(responses.compact)
-        self.parse = to_streamed_response_wrapper(responses.parse)
-        self.input_items = InputItemsWithStreamingResponse(responses.input_items)
         self.input_tokens = InputTokensWithStreamingResponse(responses.input_tokens)
 
 
@@ -530,10 +459,4 @@ class AsyncResponsesWithStreamingResponse:
     def __init__(self, responses: AsyncResponses) -> None:
         self.create = async_to_streamed_response_wrapper(responses.create)
         self.stream = async_to_streamed_response_wrapper(responses.stream)
-        self.retrieve = async_to_streamed_response_wrapper(responses.retrieve)
-        self.cancel = async_to_streamed_response_wrapper(responses.cancel)
-        self.delete = async_to_streamed_response_wrapper(responses.delete)
-        self.compact = async_to_streamed_response_wrapper(responses.compact)
-        self.parse = async_to_streamed_response_wrapper(responses.parse)
-        self.input_items = AsyncInputItemsWithStreamingResponse(responses.input_items)
         self.input_tokens = AsyncInputTokensWithStreamingResponse(responses.input_tokens)